requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-dotenv==1.0.0
pydantic==1.10.7
python-multipart==0.0.6
anthropic==0.8.1
httpx==0.24.1 